ManyLocalVars is a separate benchmark
